docs/api: document /_ping?capabilities=n endpoint
We've added a query parameter `capabilities` to the `/_ping` endpoint
that allows clients to request capabilities that the engine supports
in the `/_ping` response.

Document this change in `docs/api/version-history.md`, and update the
swagger yaml accordingly.

For Swagger, I've added a "separate" endpoint documenting the new
behavior. This has to do with Swagger not supporting different
responses for the same path (either based on query params or
content-types/other headers) – see [OAI/OpenAPI-Specification#146](https://github.com/OAI/OpenAPI-Specification/issues/146).

This is supported in OpenAPI 3.0
[(OAI/OpenAPI-Specification#146 (comment)](https://github.com/OAI/OpenAPI-Specification/issues/146#issuecomment-256648748)),
so we could fix this up in the future.

Signed-off-by: Laura Brehm <laurabrehm@hey.com>