libnetwork/drivers/bridge: remove ErrInvalidDriverConfig
It's a generic error, doesn't implement an errdefs type, is poorly formatted,
and not used as sentinel error anywhere. Let's remove it, and update the error
produced to be slightly more informative (include the invalid type). Worth
noting that there's no need to include the name of the driver in the error,
because the only uses of the error (in `registerNetworkDrivers`) already
decorates it; https://github.com/moby/moby/blob/5fd7ed2937a2308681e9e8c4b049231a84b3ca88/libnetwork/drivers_linux.go#L34-L36

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>