libnetwork/drivers/bridge: remove "InternalError()" method from errors
The `InternalError()` method was added in [moby/libnetwork@50964c9] to
classify the error. However, the same commit defined interfaces for error
types (in the types package). The [InternalError] interface defined did
not match, as it defines a `Internal()` method instead of `InternalError()`.

In short; these errors were never matching any interface, and the actual
error implementations themselves were also never used as a sentinel error,
so we can safely remove these methods.

[moby/libnetwork@50964c9]: https://github.com/moby/libnetwork/commit/50964c994831c3a960d306ff130ff42983e2fe4a
[InternalError]: https://github.com/moby/libnetwork/blob/50964c994831c3a960d306ff130ff42983e2fe4a/types/types.go#L233-L237

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>