pkg/idtools: un-deprecate Windows consts for now
These consts were deprecated in 9c368a93b690735bccfdb7c371371f6ec324d81e, but are
used externally and lack a canonical location. These sids are "special", as they
are available by default in Windows containers, but we need to;

- Reference official documentation / specification for that.
- Add names (not just the sid)
- Consider finding a canonical location for these consts, which could be as part
  of the OCI specs, or hcsshim (or otherwise).

Lacking a good place for these, let's un-deprecate them for the time being until
we decided what's the best location for these.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>