golangci-lint: ignore "nested context" (fatcontext) in tests
Commit 15fbd67407613d63f9632bb9cfa19ad8c5121c3f updated golangci-lint,
which came with an updated version of the "fatcontext" linter, causing
linting to fail.

We use these to set up spans for our tests; suppress these through the
golangci-lint config (instead of linline "//nolint" comments) so that
we can revisit this approach and see if there's a more idiomatic way.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>