daemon/cluster: remove Config.WatchStream and move to constructor
The WatchStream field was set as configuration option in cmd/dockerd,
but not configurable. Move creating the stream to the constructor,
and remove the configuration option. This field was introduced in
59d45c384a2de7bca73296ce1471646db14cb0c8, at which time the cmd/dockerd
code needed direct access to the stream, but a later refactor in
05346355dbd4300a21bfd64cda93ea6c5aef0cf0 introduced an accessor
(GetWatchStream) for this.

The cluster.Config struct is only used internally, it's unlikely
for any external project to use this, so skipping deprecation.

With this change, the cmd/dockerd package no longer has a direct
import of swarmkit.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>