registry: rewrite ParseRepositoryInfo to not depend on IndexInfo
This function was introduced in 568f86eb186731b907b659e4ec64bda21c2fe31d
to replace [ResolveRepositoryName]. The function was implemented to use
various parts of the registry package that were designed for the daemon
code, which was written with the assumption that it had registry-config
available from the daemon's configuration. However, `ParseRepositoryInfo`
was used in the client / CLI, which does not have this information.

To work around this problem, the code used a dummy "emptyServiceConfig"
to allow the `Insecure` and `Mirrors` fields to be propagated based on
the same defaults as used by the daemon. The `Mirrors` field would always
be empty, as there are no default mirrors, and (lacking access to the
daemon's config) the `Insecure` field would always default to registries
running on a loopback address (`::1/128`, `127.0.0.1/8`). It's worth noting
that neither the `Mirrors`, nor the `Insecure` field is used by the CLI.

This patch rewrites `ParseRepositoryInfo` to be self-contained, and not
depend on these constructs (and the `emptyServiceConfig`). For now, the
existing logic for `Insecure` is kept, but replaced by a simplified function
(`isInsecure`) with some optimizations for well-known loopback addresses
(`localhost`, `::1`, `127.0.0.1`) to prevent redundant DNS lookups or
parsing.

Note that similar changes should be made for [ParseSearchIndexInfo], which
has a similar fate and is also only used by the client / CLI.

[ResolveRepositoryName]: https://github.com/moby/moby/blob/11e47996dc3efe8b358f21b4ff243d9c3e5fa1ec/registry/registry.go#L199-L222
[ParseSearchIndexInfo]: https://github.com/moby/moby/blob/d86dd7594882e5c79fb039e83a2cfa065f6483c0/registry/search.go#L153-L162

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>