registry: deprecate RepositoryInfo.Official field
This field indicates whether a repository is an official image (Docker
Library / Docker Official Images). This information is deducted from the
image reference, and not used anywhere, other than in tests.

The `RepositoryInfo` and `IndexInfo` types were originally introduced in
568f86eb186731b907b659e4ec64bda21c2fe31d, with their fields documented in
4fcb9ac40ce33c4d6e08d5669af6be5e076e2574.

At the time, the `Official` field was only used for `docker push` to produce
a custom error message if someone would attempt to push an official image
to docker hub (assuming nobody would be able to do so);
https://github.com/moby/moby/blob/6870bde5847d4484d87f91baf58edc52f575ea5d/api/client/commands.go#L1184-L1194

Before that commit, the condition for this error message was based on the
given image reference directly; b370acd679622239849b1dfbe05d61804aeb7162
https://github.com/moby/moby/blob/b370acd679622239849b1dfbe05d61804aeb7162/commands.go#L421-L428

This patch deprecates the field, because it's not used, and removes
tests related to it.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>