registry: TestLoadInsecureRegistries: don't mutate emptyServiceConfig
This was not revealed in our tests, which only checked for the length
of the Mirror-slice, but when testing with DeepEquals, tests were failing
when all tests were run (but succeeded on individual tests). The problem
here is that some code can mutate the list of Mirrors and set it to `nil`
or an empty slice, resulting in other tests to fail.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>