registry: TestNewIndexInfo: assert all fields
Using DeepEquals showed that the test was missing differences between
nil-mirrors and empty-slice, in addition to mirrors being normalized
(the test only checked for the length).

While we should consider if we need an explicit empty slice (or if a
nil value would be appropriate), at least we now have a test to verify
the behavior.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>