libnetwork/resolvconf: Build: re-implement using new implementation
Re-implement the Build function using  the new implementation from
libnetwork/internal/resolvconf.

BuildKit is the only consumer of the Build function currently, and we
should consider either moving the internal package out of internal, or
to provide a more customized variant of Build that fits BuildKit's
needs (i.e., only patch content, but not write to a file).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>