layer: remove StoreOptions.MetadataStorePathTemplate
This field was added to allow templating the storage-location of the storage
driver as part of a refactor in f5916b10ae02c7db83052a97205ac345a3d96300.

In practice, the template is never customized, and always set to;
`/<data-root>/image/<driver-name>/layerdb`, where `<driver-name>` is
passed in by the driver.

This patch removes the field and its uses.

- https://github.com/moby/moby/blob/00ab386b5a2ebcf85b6a03b800f593c3a140c6a8/daemon/daemon.go#L1074-L1075
- https://github.com/moby/moby/blob/00ab386b5a2ebcf85b6a03b800f593c3a140c6a8/integration/image/remove_unix_test.go#L50-L51

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>